Move atomic related structs from api to common
