+++ common/pooll2tx.go
@@ -1,18 +1,14 @@
 package common
 
 import (
-	"database/sql/driver"
 	"encoding/binary"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"math/big"
-	"strings"
 	"time"
 
 	ethCommon "github.com/ethereum/go-ethereum/common"
-	ethCrypto "github.com/ethereum/go-ethereum/crypto"
 	"github.com/hermeznetwork/tracerr"
 	"github.com/iden3/go-iden3-crypto/babyjub"
 	"github.com/iden3/go-iden3-crypto/poseidon"
@@ -541,84 +537,5 @@ func (tx *PoolL2Tx) UnmarshalJSON(data []byte) error {
 		TokenSymbol: receivedJSON.FromIdx.TokenSymbol,
 		RqOffset:    receivedJSON.RqOffset,
 	}
-	return nil
-}
-
-// AtomicGroupIDLen is the length of a Hermez network atomic group
-const AtomicGroupIDLen = 32
-
-// AtomicGroupID is the identifier of a Hermez network atomic group
-type AtomicGroupID [AtomicGroupIDLen]byte
-
-// EmptyAtomicGroupID represents an empty Hermez network atomic group identifier
-var EmptyAtomicGroupID = AtomicGroupID([32]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-	0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0})
-
-// CalculateAtomicGroupID calculates the atomic group ID given the identifiers of
-// the transactions that conform the atomic group
-func CalculateAtomicGroupID(txIDs []TxID) AtomicGroupID {
-	txIDConcatenation := make([]byte, TxIDLen*len(txIDs))
-	for i, id := range txIDs {
-		idBytes := [TxIDLen]byte(id)
-		copy(txIDConcatenation[i*TxIDLen:(i+1)*TxIDLen], idBytes[:])
-	}
-	h := ethCrypto.Keccak256Hash(txIDConcatenation).Bytes()
-	var agid AtomicGroupID
-	copy(agid[:], h)
-	return agid
-}
-
-// Scan implements Scanner for database/sql.
-func (agid *AtomicGroupID) Scan(src interface{}) error {
-	srcB, ok := src.([]byte)
-	if !ok {
-		return tracerr.Wrap(fmt.Errorf("can't scan %T into AtomicGroupID", src))
-	}
-	if len(srcB) != AtomicGroupIDLen {
-		return tracerr.Wrap(fmt.Errorf("can't scan []byte of len %d into AtomicGroupID, need %d",
-			len(srcB), AtomicGroupIDLen))
-	}
-	copy(agid[:], srcB)
-	return nil
-}
-
-// Value implements valuer for database/sql.
-func (agid AtomicGroupID) Value() (driver.Value, error) {
-	return agid[:], nil
-}
-
-// String returns a string hexadecimal representation of the AtomicGroupID
-func (agid AtomicGroupID) String() string {
-	return "0x" + hex.EncodeToString(agid[:])
-}
-
-// NewAtomicGroupIDFromString returns a string hexadecimal representation of the AtomicGroupID
-func NewAtomicGroupIDFromString(idStr string) (AtomicGroupID, error) {
-	agid := AtomicGroupID{}
-	idStr = strings.TrimPrefix(idStr, "0x")
-	decoded, err := hex.DecodeString(idStr)
-	if err != nil {
-		return AtomicGroupID{}, tracerr.Wrap(err)
-	}
-	if len(decoded) != AtomicGroupIDLen {
-		return agid, tracerr.Wrap(errors.New("Invalid idStr"))
-	}
-	copy(agid[:], decoded)
-	return agid, nil
-}
-
-// MarshalText marshals a AtomicGroupID
-func (agid AtomicGroupID) MarshalText() ([]byte, error) {
-	return []byte(agid.String()), nil
-}
-
-// UnmarshalText unmarshalls a AtomicGroupID
-func (agid *AtomicGroupID) UnmarshalText(data []byte) error {
-	idStr := string(data)
-	id, err := NewAtomicGroupIDFromString(idStr)
-	if err != nil {
-		return tracerr.Wrap(err)
-	}
-	*agid = id
 	return nil
 }
